begin FastFit class
